+++ RELEASE_NOTES.md
@@ -4,8 +4,9 @@
 * [#693] Changes made to active effects on actors will now take effect immediately
 * [#849] The UI and functionality of the Shadowdarklings importer has been improved
 * [#853] CSS Styling for Journals
-* [#854] Added SoloDark journals
+* [#854] Added SoloDark journals, roll tables and a new oracle app
 * [#856] Beauty pass over Quickstart journals
+* [#874] Added class descriptions to the character generator and class sheet remain open during level up
 
 ## Bugfixes
 * [#852] Fixed a bug where ranged attacks were not showing correct damage
@@ -22,7 +23,6 @@
 	- Welcome message now includes a link to the video tutorial: https://youtu.be/hoBxiK71DBQ
 
 * [#872] Updated Bard abilities to the definitions in the Bard and Ranger v1.3 PDF
-* [#875] Fixes to links in Quickstart Adventure and make use of new Journal
 
 **NOTE:** This release marks the end of support for Foundry VTT v11.